requirements.txt
aiogram==3.12.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
asyncpg==0.29.0
alembic==1.13.1
psycopg[binary]

